requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
frappe-bench==5.25.3
gitdb==4.0.12
GitPython==3.1.44
honcho==2.0.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
python-crontab==2.6.0
python-dateutil==2.9.0.post0
requests==2.32.4
semantic-version==2.10.0
six==1.17.0
smmap==5.0.2
tomli==2.2.1
urllib3==2.4.0
uv==0.7.12
supabase==1.2.0
python-dotenv
reportlab
num2words
plotly
openpyxl
PyMuPDF
bcrypt
boto3
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-multipart
pandas